requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
pdfplumber==0.10.3
openai==1.12.0
python-dotenv==1.0.1